시세 코루틴 적용

+++ app/src/main/java/com/bit/kodari/Util/Upbit/UpbitResult.kt
@@ -22,7 +22,7 @@ data class UpbitPrice(
     val timestamp: Long,
     val trade_date: String,
     val trade_date_kst: String,
-    val trade_price: Int,
+    val trade_price: Double,
     val trade_time: String,
     val trade_time_kst: String,
     val trade_timestamp: Long,
